feature: packaging preparation (#12)
* style: applied black -l 100 formatting

* docs: added CONTRIBUTING.md

* refactor: renamed src into bhepop2

* fix: fixed data paths

* ci: added gh for black formatting and pytest

* fix: ci runs-on

* build: add scipy to requirements

* ci: change ubuntu version

* build: add sklearn to requirements

* build: add scikit-learn to requirements

* fix: remove an enforced variable

* docs: improved MaxEntropyEnrichment doc

* docs: improved sphinx docs

* feat: analysis of results

* refactor: move eqasim_application in notebook

* Update README.md

* fix: fixed error in main algorithm

* fix: update test data to last fix, changed method for computing probs

* docs: improve sphinx doc
