+++ bhepop2/__init__.py
@@ -1,0 +1,6 @@
+"""
+This is the main bhepop2 package.
+"""
+
+#: bhepop2 version
+__version__ = "0.0.1"
