+++ docs/about.rst
@@ -1,0 +1,23 @@
+.. _about:
+
+#############
+About Bhepop2
+#############
+
+******
+Github
+******
+
+Bhepop2 is hosted on github `here <https://github.com/tellae/bhepop2>`_.
+
+*************
+Contributions
+*************
+
+Feedback and contributions on Bhepop2, its code, its documentation, are very welcome.
+
+*******
+Authors
+*******
+
+Bhepop2 is the result of the collaboration of `UGE <https://ease.univ-gustave-eiffel.fr/>`_ and `Tellae <https://tellae.fr/>`_.
