+++ docs/contents.rst
@@ -1,0 +1,21 @@
+Documentation of Bhepop2
+========================
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   index
+   algorithm
+   inputs
+   apiref
+   about
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
