+++ docs/doc_requirements.txt
@@ -1,3 +1,4 @@
 sphinx==5.3.0
 sphinx-rtd-theme==0.4.3
 sphinx-autoapi==1.7.0
+sphinxcontrib-mermaid==0.7.1
