+++ requirements.txt
@@ -4,3 +4,5 @@ maxentropy==0.3.0
 tqdm==4.62.3
 xlrd==2.0.1
 jsonschema~=3.2.0
+scipy==1.5.1
+scikit-learn==1.1.1
