+++ tests/test_max_entropy_enrich.py
@@ -1,4 +1,4 @@
-from src.max_entropy_enrich import MaxEntropyEnrichment
+from bhepop2.max_entropy_enrich import MaxEntropyEnrichment
 from tests.conftest import *
 import numpy as np
 
@@ -8,14 +8,16 @@ def test_max_entropy_enrich():
 
     filosofi = get_filosofi_distributions()
 
-    enrich_class = MaxEntropyEnrichment(synth_pop, filosofi, CODE_INSEE, list(MODALITIES.keys()), parameters=parameters, seed=SEED)
+    enrich_class = MaxEntropyEnrichment(
+        synth_pop, filosofi, CODE_INSEE, list(MODALITIES.keys()), parameters=parameters, seed=SEED
+    )
 
     enrich_class.main()
 
     pop = enrich_class.assign_feature_value_to_pop()
 
-    # pop.to_csv("../tests/nantes_enriched.csv", index=False)
+    pop.to_csv("tests/nantes_enriched.csv", index=False)
 
-    expected_enriched_pop = pd.read_csv("../tests/nantes_enriched.csv")
+    expected_enriched_pop = pd.read_csv("tests/nantes_enriched.csv")
 
     assert np.all((pop == expected_enriched_pop).to_numpy())
